Update gulp build tools for Typescript
