+++ src/templates/componentSampleTemplate.html
@@ -11,7 +11,8 @@
   <link rel="stylesheet" href="<%= componentName %>.css">
 
   <script type="text/javascript" src="https://code.jquery.com/jquery-2.1.4.min.js"></script>
- 
+  <script type="text/javascript" src="../../../js/fabric.js"></script>
+
   <script type="text/javascript">
     $(document).ready(function() {
        // Vanilla JS Components
