+++ src/templates/individual-component-example.html
@@ -11,8 +11,9 @@
   <link rel="stylesheet" href="<%= componentName %>.css">
 
   <script type="text/javascript" src="https://code.jquery.com/jquery-2.1.4.min.js"></script>
+  <script type="text/javascript" src="../../../js/fabric.js"></script>
   
-  <%= jstag %>
+    <%= jstag %>
  
   <script type="text/javascript">
     $(document).ready(function() {
